requirements.txt
fastapi==0.81.0
uvicorn[standard]==0.18.3
opentelemetry-exporter-jaeger==1.12.0
# jaeger-client==4.8.0
opentelemetry-propagator-b3==1.12.0